libs/core/wrappers/javascript/src/program.spec.ts
import { Program, MovingAverage, PeakDetector, Connection } from "@rtbot/api";

describe("Program", () => {
  let program: Program;
  const title = "my program title";
  const description = "description";

  beforeEach(() => {
    program = new Program(title, description);
    const op1 = new MovingAverage("ma1", 2);
    const op2 = new MovingAverage("ma2", 20);
    program.addOperator(op1);
    program.addOperator(op2);
    program.addConnection(op1, op2);
  });

  it("can create a new instance", () => {
    expect(program.title).toBe(title);
  });

  it("can be serialized", () => {
    const json = JSON.stringify(program);
    const parsedProgram = JSON.parse(json);
    expect(parsedProgram.operators.find((op: any) => op.id === "ma1")).toMatchObject({ id: "ma1", n: 2 });
  });

  it("validates operator input", () => {
    const wrap = () => {
      new MovingAverage("ma", 0.5);
    };
    expect(wrap).toThrow(/(.*)invalid_type(.*)/);
    expect(wrap).toThrow(/(.*)Expected integer, received float(.*)/);
  });

  it("validates the entire program", () => {
    const wrap = () => {
      program.validate();
    };
    expect(wrap).not.toThrow();
  });
});
